Refactoring and cleaning up the code.

+++ etsin/service/fairdata-etsin-init.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 #########################################################
-# This file is the startup file for the Etsin service,
-# it does contain also those initializations
+# This file is the init file for the Etsin service
 #
 # Author(s):
 #      Juhapekka Piiroinen <juhapekka.piiroinen@csc.fi>
@@ -17,60 +16,68 @@ echo
 echo "Waiting until metax-api server is ready.."
 set +e
 while true; do
-    nc ${APP_METAX_API_HOST} 443 -z -w 59 2> /dev/null
+    nc metax.csc.local 443 -z -w 59 2> /dev/null
     if [[ $? == 0 ]]; then
-        curl https://${APP_METAX_API_HOST}/rest/datasets --insecure -f -s > /dev/null
+        curl https://metax.csc.local/rest/datasets --insecure -f -s > /dev/null
         if [[ $? == 0 ]]; then
             break
         else
             sleep 1
         fi
+    else
+        sleep 1
     fi
 done
 set -e
 echo "..ok"
 echo
 
-echo "Provisioning.."
-if [[ ! -f /usr/local/etsin/search_scripts/reindex_by_recreating_index.sh ]]; then
-    pushd /code/etsin-ops/ansible
-        ansible-galaxy install --roles-path=roles -r requirements.yml
-        ansible-playbook -i inventories/local_development provision_dataservers.yml
-        ansible-playbook -i inventories/local_development provision_webservers.yml
+
+#########################################################
+# Do the first time start up related initializations and system
+# preparing which will take care that the database itself
+# has the required databases and schemas.
+if [[ -f /first-time-init ]]; then
+
+    echo "Provisioning.."
+    if [[ ! -f /usr/local/etsin/search_scripts/reindex_by_recreating_index.sh ]]; then
+        pushd /code/etsin-ops/ansible
+            ansible-galaxy install --roles-path=roles -r requirements.yml
+            ansible-playbook -i inventories/local_development provision_dataservers.yml
+            ansible-playbook -i inventories/local_development provision_webservers.yml
+        popd
+    fi
+    echo "..ok"
+    echo
+
+    echo "Configuring.."
+
+    systemctl enable memcached
+    systemctl enable nginx
+    systemctl enable rabbitmq-consumer
+
+    systemctl start memcached
+    systemctl start nginx
+    systemctl start rabbitmq-consumer
+
+    pushd /home/etsin-user
+        test -f /root/advanced_settings.json || ln -s /home/etsin-user/advanced_settings.json /root/advanced_settings.json
+        test -f /root/app_config || ln -s /home/etsin-user/app_config /root/app_config
+        test -f /root/settings.json || ln -s /home/etsin-user/settings.json /root/settings.json
+        chmod +r /home/etsin-user/advanced_settings.json /home/etsin-user/app_config /home/etsin-user/settings.json
     popd
+    echo "..configure done."
+    echo
+
+    echo "Reindexing.."
+    su - etsin-user -c "/usr/local/etsin/search_scripts/reindex_by_recreating_index.sh"
+    echo "..reindex done."
+    echo
+
+    echo "Marking init as complete.."
+    rm -f /first-time-init
+    echo "..marked."
+    echo
 fi
-echo "..ok"
-echo
 
-echo "Configuring.."
-rm -f /first-time-init
-
-systemctl enable memcached
-systemctl enable nginx
-systemctl enable rabbitmq-consumer
-
-systemctl start memcached
-systemctl start nginx
-systemctl start rabbitmq-consumer
-
-pushd /home/etsin-user
-    test -f /root/advanced_settings.json || ln -s /home/etsin-user/advanced_settings.json /root/advanced_settings.json
-    test -f /root/app_config || ln -s /home/etsin-user/app_config /root/app_config
-    test -f /root/settings.json || ln -s /home/etsin-user/settings.json /root/settings.json
-    chmod +r /home/etsin-user/advanced_settings.json /home/etsin-user/app_config /home/etsin-user/settings.json
-popd
-echo "..configure done."
-echo
-
-echo "Reindexing.."
-su - etsin-user -c "/usr/local/etsin/search_scripts/reindex_by_recreating_index.sh"
-echo "..reindex done."
-echo
-
-echo "Starting etsin_finder.."
-pushd /code/etsin_finder
-    /usr/local/etsin/pyenv/bin/gunicorn --bind unix:/usr/local/etsin/gunicorn/socket --access-logfile - --error-logfile - --config /etc/gunicorn.py --reload etsin_finder.finder:app
-popd
-echo "..done."
-echo
 exit 0
